Tiny Grenzelhoft Changes (Sidearms, New Class, and Better RNG System)
Grenzelhoft sweep 2024

## About The Pull Request

Fixes up the old loadout system for Grenzlehoft mercs, gives them sidearms, and adds a third, heavier loadout.

## Why It's Good For The Game

Grenzlehoft Mercenaries are currently mediocre compared to Desert Riders, and need a facelift. The new legally-distinct not-a-landsknecht is meant to act as a leader for the group and for someone for the mercs to take cover behind. Their armor is a bit better, but they have no leg protection roundstart. Sidearms also help fix a problem that grenzels had with their old loadouts, where polearm rng meant that luck could make you significantly worse than your fellow mercs, now you have a messer to work with.
Seems to be throwing an error that matches the one in the eye pr, not sure if it can be fixed on my end.